# Backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_video_api_v1_videos_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_video_api_v1_videos_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RouteType": {
        "enum": [
          "emergency",
          "housing_stabilization",
          "return_relocation"
        ],
        "title": "RouteType",
        "type": "string"
      },
      "SourceCreate": {
        "properties": {
          "contact": {
            "minLength": 2,
            "title": "Contact",
            "type": "string"
          },
          "coverage": {
            "maxLength": 180,
            "minLength": 2,
            "title": "Coverage",
            "type": "string"
          },
          "entity": {
            "maxLength": 180,
            "minLength": 2,
            "title": "Entity",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Id",
            "type": "string"
          },
          "program": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Program",
            "type": "string"
          },
          "requirements": {
            "minLength": 2,
            "title": "Requirements",
            "type": "string"
          },
          "route_types": {
            "items": {
              "$ref": "#/components/schemas/RouteType"
            },
            "minItems": 1,
            "title": "Route Types",
            "type": "array"
          },
          "source_kind": {
            "enum": [
              "contact",
              "program"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "url": {
            "pattern": "^https://",
            "title": "Url",
            "type": "string"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "entity",
          "program",
          "coverage",
          "requirements",
          "contact",
          "url",
          "verified_at",
          "expires_at",
          "source_kind",
          "route_types"
        ],
        "title": "SourceCreate",
        "type": "object"
      },
      "SourceList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SourceList",
        "type": "object"
      },
      "SourceRead": {
        "properties": {
          "contact": {
            "title": "Contact",
            "type": "string"
          },
          "coverage": {
            "title": "Coverage",
            "type": "string"
          },
          "disclaimer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disclaimer"
          },
          "entity": {
            "title": "Entity",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_expired": {
            "title": "Is Expired",
            "type": "boolean"
          },
          "program": {
            "title": "Program",
            "type": "string"
          },
          "requirements": {
            "title": "Requirements",
            "type": "string"
          },
          "route_types": {
            "items": {
              "$ref": "#/components/schemas/RouteType"
            },
            "title": "Route Types",
            "type": "array"
          },
          "source_kind": {
            "enum": [
              "contact",
              "program"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "entity",
          "program",
          "coverage",
          "requirements",
          "contact",
          "url",
          "verified_at",
          "expires_at",
          "source_kind",
          "route_types",
          "status",
          "is_expired",
          "disclaimer"
        ],
        "title": "SourceRead",
        "type": "object"
      },
      "SourceUpdate": {
        "properties": {
          "contact": {
            "anyOf": [
              {
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "coverage": {
            "anyOf": [
              {
                "maxLength": 180,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "entity": {
            "anyOf": [
              {
                "maxLength": 180,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "program": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program"
          },
          "requirements": {
            "anyOf": [
              {
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirements"
          },
          "route_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RouteType"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Types"
          },
          "source_kind": {
            "anyOf": [
              {
                "enum": [
                  "contact",
                  "program"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Kind"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "url": {
            "anyOf": [
              {
                "pattern": "^https://",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "title": "SourceUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VideoRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_demo": {
            "title": "Is Demo",
            "type": "boolean"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "media_type",
          "size_bytes",
          "status",
          "is_demo",
          "created_at"
        ],
        "title": "VideoRead",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "SENDA API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sources": {
      "get": {
        "operationId": "list_sources_api_v1_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Sources",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "operationId": "create_source_api_v1_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Source",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/sources/{source_id}": {
      "delete": {
        "operationId": "retire_source_api_v1_sources__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Retire Source",
        "tags": [
          "administration"
        ]
      },
      "get": {
        "operationId": "get_source_api_v1_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Source",
        "tags": [
          "administration"
        ]
      },
      "patch": {
        "operationId": "update_source_api_v1_sources__source_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Source",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/api/v1/videos": {
      "post": {
        "operationId": "upload_video_api_v1_videos_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video_api_v1_videos_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Upload Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/v1/videos/demo": {
      "post": {
        "operationId": "create_demo_video_api_v1_videos_demo_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Demo Video",
        "tags": [
          "videos"
        ]
      }
    },
...
    "/api/v1/videos/{video_id}/stream": {
      "get": {
        "operationId": "stream_video_api_v1_videos__video_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Stream Video",
        "tags": [
          "videos"
        ]
      }
    }
  }
}